Added GetAllProducts() to Index method in ProductController, Created Product navigation button on Home page

+++ SuggestedOrdering/Models/Product.cs
@@ -11,7 +11,7 @@ namespace SuggestedOrdering.Models
         public int ItemNumber { get; set; }
         public string Description { get; set; }
         public string CaseSize { get; set; }
-        public int UsageNumber { get; set; }
-        public int OnHand { get; set; }
+        public double UsageNumber { get; set; }
+        public double OnHand { get; set; }
     }
 }
